+++ src/Config.hs
@@ -4,6 +4,8 @@ module Config
     , defaultConfig
     ) where
 
+import System.IO (FilePath)
+
 data IOMode = Chars | Nums
     deriving (Eq, Show)
 
@@ -11,6 +13,7 @@ data Config = Config
     { inputMode :: IOMode
     , outputMode :: IOMode
     , isDebug :: Bool
+    , source :: Maybe FilePath
     }
 
 defaultConfig :: Config
@@ -18,4 +21,5 @@ defaultConfig = Config
     { inputMode = Chars
     , outputMode = Chars
     , isDebug = False
+    , source = Nothing
     }